update backend & remove old types
