+++ types/index.ts
@@ -1,4 +1,4 @@
-export * from "./responses";
-// alias this from @aws-amplify/storage
-export type StorageUploadDataPayload = Blob | ArrayBufferView | ArrayBuffer | string;
-
+export * from "../graphql/API";
+// i think all the old types are, well, old.
+// regenerated the graphql schema and that's in ../graphql/API.ts mostly
+// TODO: add other types here
